updated project files

+++ LearningSwiftCourse/LearningSwiftCourse/MVC/Views/CirclePieView.swift
@@ -7,6 +7,7 @@
 //
 
 import UIKit
+import LearningSwiftCourseExtensions
 
 @IBDesignable 
 class CirclePieView: UIView {
@@ -96,7 +97,7 @@ class CirclePieView: UIView {
             let midPath = UIBezierPath()
             midPath.move(to: viewCenter)
             
-            midPath.addArc(withCenter: viewCenter, radius: CGFloat(radius), startAngle: startAngle.degreesToRadians, endAngle: endAngle.degreesToRadians, clockwise: true)
+            midPath.addArc(withCenter: viewCenter, radius: CGFloat(radius), startAngle: CGFloat(startAngle.degreesToRadians), endAngle: CGFloat(endAngle.degreesToRadians), clockwise: true)
             
             midPath.close()
             midPath.fill()
